ready lab9

+++ src/main/java/ua/lviv/iot/gameroom/repositories/ChildRepository.java
@@ -3,5 +3,5 @@ package ua.lviv.iot.gameroom.repositories;
 import org.springframework.data.repository.CrudRepository;
 import ua.lviv.iot.gameroom.models.Child;
 
-public interface ChildRepository extends CrudRepository<Child, Integer> {
+public interface ChildRepository extends CrudRepository<Child, Long> {
 }
